+++ client/packages/portal-core/src/index.ts
@@ -1,7 +1,7 @@
 export * from './create-portal-framework/create-portal-framework';
-export * from './hooks/use-app-loader';
-export * from './hooks/use-phases';
+
 export * from './types/portal-config';
 export * from './menu';
 export * from './avatar';
 export * from './phases';
+export * from './hooks';
